+++ apps/web/server/pages.ts
@@ -1,0 +1,6 @@
+export const pages = {
+  signIn: "/signin",
+  signOut: "/signin",
+  error: "/signin",
+  newUser: "/signin",
+} as const;
